Rename 'rocketmq-client-java' to 'rocketmq-client'
