--- instrumentation/rocketmq/rocketmq-client/rocketmq-client-java-5.0/javaagent/build.gradle.kts
@@ -1,18 +1,0 @@
-plugins {
-  id("otel.javaagent-instrumentation")
-}
-
-muzzle {
-  pass {
-    group.set("org.apache.rocketmq")
-    module.set("rocketmq-client-java")
-    versions.set("[5.0.0,)")
-    assertInverse.set(true)
-  }
-}
-
-dependencies {
-  library("org.apache.rocketmq:rocketmq-client-java:5.0.0")
-
-  testImplementation(project(":instrumentation:rocketmq:rocketmq-client:rocketmq-client-java-5.0:testing"))
-}
