--- instrumentation/rocketmq/rocketmq-client/rocketmq-client-java-5.0/javaagent/src/test/java/io/opentelemetry/instrumentation/rocketmqclientjava/v5_0/RocketMqClientTest.java
@@ -1,20 +1,0 @@
-/*
- * Copyright The OpenTelemetry Authors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package io.opentelemetry.instrumentation.rocketmqclientjava.v5_0;
-
-import io.opentelemetry.instrumentation.testing.junit.AgentInstrumentationExtension;
-import io.opentelemetry.instrumentation.testing.junit.InstrumentationExtension;
-import org.junit.jupiter.api.extension.RegisterExtension;
-
-public class RocketMqClientTest extends AbstractRocketMqClientTest {
-  @RegisterExtension
-  static final InstrumentationExtension testing = AgentInstrumentationExtension.create();
-
-  @Override
-  protected InstrumentationExtension testing() {
-    return testing;
-  }
-}
