+++ settings.gradle.kts
@@ -404,8 +404,8 @@ include(":instrumentation:rmi:javaagent")
 include(":instrumentation:rocketmq:rocketmq-client:rocketmq-client-4.8:javaagent")
 include(":instrumentation:rocketmq:rocketmq-client:rocketmq-client-4.8:library")
 include(":instrumentation:rocketmq:rocketmq-client:rocketmq-client-4.8:testing")
-include(":instrumentation:rocketmq:rocketmq-client:rocketmq-client-java-5.0:javaagent")
-include(":instrumentation:rocketmq:rocketmq-client:rocketmq-client-java-5.0:testing")
+include(":instrumentation:rocketmq:rocketmq-client:rocketmq-client-5.0:javaagent")
+include(":instrumentation:rocketmq:rocketmq-client:rocketmq-client-5.0:testing")
 include(":instrumentation:runtime-metrics:javaagent")
 include(":instrumentation:runtime-metrics:library")
 include(":instrumentation:rxjava:rxjava-1.0:library")
